Display task can set two pixel successful
