+++ app/display/graphic.h
@@ -1,0 +1,4 @@
+#ifndef GRAPHIC_H
+#define GRAPHIC_H
+
+#endif
