+++ app/display/oled.h
@@ -9,14 +9,14 @@ typedef enum oled_status {
 
 oled_status_t myoledstatus;
 
-
 #define SSD1306 1
 #define SSH1106 1
 
 #define SSD1306_address 0x3C
 
 //Defines for later use
-#define SSD1306_LCDWIDTH                  128
+#define SSD1306_LCDWIDTH            128
+#define SSD1306_LCDHEIGHT           64
 
 //  Fundamental #defines
 #define SSD1306_SETCONTRAST 0x81
@@ -63,5 +63,7 @@ oled_status_t myoledstatus;
 
 #define SSD1306_EXTERNALVCC 0x1
 #define SSD1306_SWITCHCAPVCC 0x2
+
+uint8_t SSD1306_buffer[SSD1306_LCDWIDTH * SSD1306_LCDHEIGHT / 8];
 
 #endif
