+++ app/user/user_main.c
@@ -1,6 +1,6 @@
 #include "esp_common.h"
 #include "user_config.h"
-#include "../oled/oled.h"
+//#include "../display/display.h"
 
 uint32 user_rf_cal_sector_set(void)
 {
@@ -43,7 +43,7 @@ void ICACHE_FLASH_ATTR wifi_event_handler_cb(System_Event_t *event)
             printf("Disconnected!\n");
             break;
         case EVENT_STAMODE_CONNECTED:
-            myoledstatus = oled_display_cloud;
+            //myoledstatus = oled_display_cloud;
             printf("Connected!\n");
             break;
         case EVENT_STAMODE_GOT_IP:
@@ -61,8 +61,9 @@ void user_init(void)
 
     os_printf("SDK version:%s\n", system_get_sdk_version());
 
-    myoledstatus = oled_connecting_wifi;
-    oled_start();
+    //myoledstatus = oled_connecting_wifi;
+    //oled_start();
+    display_start();
 
     wifi_set_opmode(STATION_MODE);
